v0.3
Code is at build stage but I lied as it is not finalised, however I am
trying to find a solution to calling main() within the UserRestart()
function so I don't have to rewrite the restart per each thing. I am
going to discuss this with Will.

Release v0.4 will be the bug-free release then v1.0 after that will be
the cleaned and final release